feat: wait for mkcert to create root ca

+++ install.sh
@@ -3,10 +3,16 @@
 git clone https://github.com/saul-salazar-dotcom/homelabx.net.git ~/homelabx
 
 cd ~/homelabx/proxy && docker compose up -d
-echo "⏳ Wait while mkcert creates a root CA"
-sleep 5
-ln -s proxy/rootCA.pem plan/rootCA.pem
-ln -s proxy/rootCA.pem design/rootCA.pem
 
-cd ~/homelabx
-./start.sh
+FILE="$HOME/homelabx/proxy/rootCA.pem"
+echo "⏳ Waiting for mkcert to create a root CA..."
+while true; do
+    if [[ -f "$FILE" ]]; then
+        ln -s proxy/rootCA.pem plan/rootCA.pem
+        ln -s proxy/rootCA.pem design/rootCA.pem
+        break
+    fi
+    sleep 5  # Wait for 5 seconds before checking again
+done
+
+cd ~/homelabx && ./start.sh
